# spec/openapi.yaml
openapi: "3.1.0"
info:
  title: "标题：ArchForge接口文档"
  description: "描述：文档说明"
  contact:
    name: "ArchForge"
  version: "版本号:1.0.0"
  license:
    name: "Apache-2.0"
    identifier: "Apache-2.0"
servers:
- url: "/"
security:
- bearer: []
...
paths:
...
  /admin/blog/file/upload:
    post:
      tags:
      - "blog-file-controller"
      operationId: "uploadMarkdownImage"
      parameters:
      - name: "file[]"
        in: "query"
        required: true
        schema:
          type: "array"
          items:
            type: "string"
            format: "binary"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "object"
                additionalProperties:
                  type: "object"
        "4XX":
          $ref: "#/components/responses/Error"
      summary: "Upload markdown image"
...
  /admin/file:
    get:
      tags:
      - "文件管理"
      summary: "获取文件列表"
      operationId: "listFiles"
      parameters:
      - name: "params"
        in: "query"
        required: true
        schema:
          $ref: "#/components/schemas/FileListRequest"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPageResponseFileResponse"
        "4XX":
          $ref: "#/components/responses/Error"
  /admin/file/download/{fileId}:
    get:
      tags:
      - "文件管理"
      summary: "下载文件"
      operationId: "downloadFile"
      parameters:
      - name: "fileId"
        in: "path"
        required: true
        schema:
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "string"
                format: "binary"
        "4XX":
          $ref: "#/components/responses/Error"
  /admin/file/upload:
    post:
      tags:
      - "文件管理"
      summary: "上传文件"
      operationId: "uploadFile"
      requestBody:
        content:
          application/json:
            schema:
              required:
              - "file"
              type: "object"
              properties:
                file:
                  type: "string"
                  format: "binary"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadFileResponse"
        "4XX":
          $ref: "#/components/responses/Error"
  /admin/file/upload-image:
    post:
      tags:
      - "文件管理"
      summary: "上传图片（头像等）"
      operationId: "uploadImage"
      requestBody:
        content:
          application/json:
            schema:
              required:
              - "file"
              type: "object"
              properties:
                file:
                  type: "string"
                  format: "binary"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadFileResponse"
        "4XX":
          $ref: "#/components/responses/Error"
  /admin/file/{fileId}:
    delete:
      tags:
      - "文件管理"
      summary: "删除文件"
      operationId: "deleteFile"
      parameters:
      - name: "fileId"
        in: "path"
        required: true
        schema:
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "boolean"
        "4XX":
          $ref: "#/components/responses/Error"
...
  /web/file/upload:
    post:
      tags:
      - "web-file-controller"
      operationId: "upload"
      requestBody:
        content:
          application/json:
            schema:
              type: "object"
              properties:
                file:
                  type: "string"
                  format: "binary"
              required:
              - "file"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/FileUploadResponse"
        "4XX":
          $ref: "#/components/responses/Error"
      summary: "Upload"
  /web/file/{fileId}:
    get:
      tags:
      - "web-file-controller"
      operationId: "download"
      parameters:
      - name: "fileId"
        in: "path"
        required: true
        schema:
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "string"
                format: "binary"
        "4XX":
          $ref: "#/components/responses/Error"
      summary: "Download"
...
components:
  schemas:
...
    AdminPageResponseFileResponse:
      type: "object"
      properties:
        list:
          type: "array"
          items:
            $ref: "#/components/schemas/FileResponse"
        total:
          type: "integer"
          format: "int64"
        pageSize:
          type: "integer"
          format: "int32"
        currentPage:
          type: "integer"
          format: "int32"
...
    FileListRequest:
      type: "object"
      properties:
        currentPage:
          type: "integer"
          format: "int32"
        pageSize:
          type: "integer"
          format: "int32"
        originalName:
          type: "string"
        storageType:
          type: "string"
    FileResponse:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        originalName:
          type: "string"
        storageName:
          type: "string"
        storagePath:
          type: "string"
        fileSize:
          type: "integer"
          format: "int64"
        contentType:
          type: "string"
        extension:
          type: "string"
        storageType:
          type: "string"
        createTime:
          type: "integer"
          format: "int64"
    FileUploadResponse:
      type: "object"
      properties:
        fileId:
          type: "integer"
          format: "int64"
        url:
          type: "string"
        name:
          type: "string"
...
    UploadFileResponse:
      type: "object"
      properties:
        fileId:
          type: "integer"
          format: "int64"
        originalName:
          type: "string"
        url:
          type: "string"
        fileSize:
          type: "integer"
          format: "int64"
...
  securitySchemes:
    bearer:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
  responses:
    Error:
      description: "Error response (RFC 9457 ProblemDetail)"
      content:
        application/json:
          schema:
            type: "object"
            properties:
              type:
                type: "string"
              title:
                type: "string"
              status:
                type: "integer"
              detail:
                type: "string"
              instance:
                type: "string"